refactor: rename `patchset_details` to `render_patchset`
